Fix global schema lookup on Windows
When using the glob match `./dir/*.md` on Windows, it gets transformed to `dir\\*.md` by the path normalization.

minimatch specifically says not to use `\` or `\\` as these are easy to mess up escaping, and the above transformation does not match when you would expect it to.

The solution is to always use `/` for the glob, or to ignore the escapes on Windows with the `windowsPathsNoEscape` option. https://github.com/isaacs/minimatch?tab=readme-ov-file#windows

Given the normalization is needed for sorting out relative paths, using the `windowsPathsNoEscape` option makes most sense. I've tested this locally.